requirements.txt
altgraph                        0.17.3
appdirs                         1.4.4
async-generator                 1.10
attrs                           22.2.0
autopep8                        2.0.2
certifi                         2022.12.7
cffi                            1.15.1
chardet                         5.1.0
comtypes                        1.2.0
et-xmlfile                      1.1.0
exceptiongroup                  1.1.1
Faker                           18.2.0
h11                             0.14.0
idna                            3.4
lml                             0.1.0
numpy                           1.25.0
openpyxl                        3.1.2
outcome                         1.2.0
pefile                          2023.2.7
pip                             23.2.1
pycodestyle                     2.10.0
pycparser                       2.21
pyee                            8.1.0
pyexcel                         0.7.0
pyexcel-io                      0.6.6
pyinstaller                     5.9.0
pyinstaller-hooks-contrib       2023.1
PySocks                         1.7.1
python-dateutil                 2.8.2
pywin32                         306
pywin32-ctypes                  0.2.0
pywinauto                       0.6.8
PyYAML                          6.0
robotframework                  6.0.2
robotframework-faker            5.0.0
robotframework-pythonlibcore    4.1.2
robotframework-selenium2library 3.0.0
robotframework-seleniumlibrary  6.0.0
scipy                           1.10.1
selenium                        4.8.2
setuptools                      65.5.0
six                             1.16.0
sniffio                         1.3.0
sortedcontainers                2.4.0
sparse-dot-topn                 0.3.4
texttable                       1.6.7
trio                            0.22.0
trio-websocket                  0.10.2
urllib3                         1.26.15
websockets                      10.1
wrapt                           1.15.0
wsproto                         1.2.0
xlrd                            2.0.1
zipp                            3.15.0